change last_modified to BigInteger

+++ base/models.py
@@ -42,7 +42,7 @@ class FileContributors(models.Model):
     file_path = models.TextField()
     file_path_hash = models.CharField(max_length=12)
 
-    last_modified = models.IntegerField()
+    last_modified = models.BigIntegerField()
 
     # email addresses seperated by comma
     emails = models.TextField()
